+++ src/Utils/StringUtils.h
@@ -10,7 +10,7 @@ namespace Raycer
 	public:
 
 		static bool endsWith(const std::string& input, const std::string& end);
-		static bool readUntilSpace(const std::string& input, int& startIndex, std::string& result);
+		static bool readUntilSpace(const std::string& input, size_t& startIndex, std::string& result);
 		static double parseDouble(const std::string& input);
 	};
 }
